Handle oddly sized disk images
The bootloader for ARM64 is written as a raw .bin file, so it can be any
size. This was not handled correctly by the current image_aggregator
code which assumed the images were all multiples of at least 512b. This
caused some rounding errors which truncated the bootloader partition.

Fix this by rounding up the sizes of the partitions to a multiple of 4k,
which due to the alignment of the GPT header also aligns them to 4k.
This fixes the 512b alignment issue and preserves the performance
niceties of 4k alignment.

Also, switch to using AlignToPowerOf2() and update the disk size
alignment code for readability.

Bug: 170433295

+++ cvd/common/libs/utils/size_utils.h
@@ -20,6 +20,6 @@
 namespace cuttlefish {
 
 // Returns the smallest multiple of 2^align_log greater than or equal to val.
-uint32_t AlignToPowerOf2(uint32_t val, uint8_t align_log);
+uint64_t AlignToPowerOf2(uint64_t val, uint8_t align_log);
 
 }  // namespace cuttlefish
